Enable setting host on a TChannel listener.
This feature is primarily for consumers who would like to explicitly set
the hostname of the listener, or for those who want to listen on a
Loopback address.
